requirements.txt

pydantic==1.10.8
requests==2.28.1
superagi-tools==1.0.8
                